feat: split vitest into separate packages
