+++ packages/expect/src/constants.ts
@@ -1,3 +1,3 @@
-export const GLOBAL_EXPECT = Symbol.for('expect-global')
 export const MATCHERS_OBJECT = Symbol.for('matchers-object')
 export const JEST_MATCHERS_OBJECT = Symbol.for('$$jest-matchers-object')
+export const GLOBAL_EXPECT = Symbol.for('expect-global')
