+++ packages/expect/src/jest-matcher-utils.ts
@@ -1,35 +1,14 @@
-// we are using only the ones needed by @testing-library/jest-dom
-// if you need more, just ask
-
 import c from 'picocolors'
-import type { PrettyFormatOptions } from 'pretty-format'
-import { format as prettyFormat, plugins as prettyFormatPlugins } from 'pretty-format'
-import { unifiedDiff } from '../../utils/diff'
-import type { DiffOptions, MatcherHintOptions } from '../../types/matcher-utils'
+import { stringify, unifiedDiff } from '@vitest/utils'
+import type { DiffOptions, MatcherHintOptions } from './types'
+
+export { stringify }
 
 export const EXPECTED_COLOR = c.green
 export const RECEIVED_COLOR = c.red
 export const INVERTED_COLOR = c.inverse
 export const BOLD_WEIGHT = c.bold
 export const DIM_COLOR = c.dim
-
-const {
-  AsymmetricMatcher,
-  DOMCollection,
-  DOMElement,
-  Immutable,
-  ReactElement,
-  ReactTestComponent,
-} = prettyFormatPlugins
-
-const PLUGINS = [
-  ReactTestComponent,
-  ReactElement,
-  DOMElement,
-  DOMCollection,
-  Immutable,
-  AsymmetricMatcher,
-]
 
 export function matcherHint(
   matcherName: string,
@@ -39,12 +18,12 @@ export function matcherHint(
 ) {
   const {
     comment = '',
-    expectedColor = EXPECTED_COLOR,
     isDirectExpectCall = false, // seems redundant with received === ''
     isNot = false,
     promise = '',
-    receivedColor = RECEIVED_COLOR,
     secondArgument = '',
+    expectedColor = EXPECTED_COLOR,
+    receivedColor = RECEIVED_COLOR,
     secondArgumentColor = EXPECTED_COLOR,
   } = options
   let hint = ''
@@ -101,35 +80,6 @@ const SPACE_SYMBOL = '\u{00B7}' // middle dot
 // replace common spaces with middle dot at the end of any line.
 const replaceTrailingSpaces = (text: string): string =>
   text.replace(/\s+$/gm, spaces => SPACE_SYMBOL.repeat(spaces.length))
-
-export function stringify(object: unknown, maxDepth = 10, { maxLength, ...options }: PrettyFormatOptions & { maxLength?: number } = {}): string {
-  const MAX_LENGTH = maxLength ?? 10000
-  let result
-
-  try {
-    result = prettyFormat(object, {
-      maxDepth,
-      escapeString: false,
-      // min: true,
-      plugins: PLUGINS,
-      ...options,
-    })
-  }
-  catch {
-    result = prettyFormat(object, {
-      callToJSON: false,
-      maxDepth,
-      escapeString: false,
-      // min: true,
-      plugins: PLUGINS,
-      ...options,
-    })
-  }
-
-  return result.length >= MAX_LENGTH && maxDepth > 1
-    ? stringify(object, Math.floor(maxDepth / 2))
-    : result
-}
 
 export const printReceived = (object: unknown): string =>
   RECEIVED_COLOR(replaceTrailingSpaces(stringify(object)))
