+++ packages/spy/rollup.config.js
@@ -1,0 +1,51 @@
+import { builtinModules } from 'module'
+import esbuild from 'rollup-plugin-esbuild'
+import dts from 'rollup-plugin-dts'
+import { defineConfig } from 'rollup'
+import pkg from './package.json'
+
+const external = [
+  ...builtinModules,
+  ...Object.keys(pkg.dependencies || {}),
+  ...Object.keys(pkg.peerDependencies || {}),
+]
+
+const plugins = [
+  esbuild({
+    target: 'node14',
+  }),
+]
+
+export default defineConfig([
+  {
+    input: 'src/index.ts',
+    output: {
+      dir: 'dist',
+      format: 'esm',
+      entryFileNames: '[name].mjs',
+      chunkFileNames: 'chunk-[name].mjs',
+    },
+    external,
+    plugins,
+    onwarn,
+  },
+  {
+    input: 'src/index.ts',
+    output: {
+      dir: 'dist',
+      entryFileNames: '[name].d.ts',
+      format: 'esm',
+    },
+    external,
+    plugins: [
+      dts({ respectExternal: true }),
+    ],
+    onwarn,
+  },
+])
+
+function onwarn(message) {
+  if (['EMPTY_BUNDLE', 'CIRCULAR_DEPENDENCY'].includes(message.code))
+    return
+  console.error(message)
+}
