+++ packages/utils/rollup.config.js
@@ -1,0 +1,64 @@
+import { builtinModules } from 'module'
+import { defineConfig } from 'rollup'
+import esbuild from 'rollup-plugin-esbuild'
+import dts from 'rollup-plugin-dts'
+import resolve from '@rollup/plugin-node-resolve'
+import json from '@rollup/plugin-json'
+import pkg from './package.json'
+
+const entries = {
+  index: 'src/index.ts',
+  helpers: 'src/helpers.ts',
+  diff: 'src/diff.ts',
+  types: 'src/types.ts',
+}
+
+const external = [
+  ...builtinModules,
+  ...Object.keys(pkg.dependencies || {}),
+  ...Object.keys(pkg.peerDependencies || {}),
+]
+
+const plugins = [
+  resolve({
+    preferBuiltins: true,
+  }),
+  json(),
+  esbuild({
+    target: 'node14',
+  }),
+]
+
+export default defineConfig([
+  {
+    input: entries,
+    output: {
+      dir: 'dist',
+      format: 'esm',
+      entryFileNames: '[name].mjs',
+      chunkFileNames: 'chunk-[name].mjs',
+    },
+    external,
+    plugins,
+    onwarn,
+  },
+  {
+    input: entries,
+    output: {
+      dir: 'dist',
+      entryFileNames: '[name].d.ts',
+      format: 'esm',
+    },
+    external,
+    plugins: [
+      dts({ respectExternal: true }),
+    ],
+    onwarn,
+  },
+])
+
+function onwarn(message) {
+  if (['EMPTY_BUNDLE', 'CIRCULAR_DEPENDENCY'].includes(message.code))
+    return
+  console.error(message)
+}
