+++ packages/utils/src/helpers.ts
@@ -1,0 +1,10 @@
+export function assertTypes(value: unknown, name: string, types: string[]): void {
+  const receivedType = typeof value
+  const pass = types.includes(receivedType)
+  if (!pass)
+    throw new TypeError(`${name} value must be ${types.join(' or ')}, received "${receivedType}"`)
+}
+
+export function isObject(item: unknown): boolean {
+  return item != null && typeof item === 'object' && !Array.isArray(item)
+}
