+++ packages/vitest/package.json
@@ -132,6 +132,9 @@
     "@types/prompts": "^2.4.2",
     "@types/sinonjs__fake-timers": "^8.1.2",
     "@vitest/ui": "workspace:*",
+    "@vitest/expect": "workspace:*",
+    "@vitest/spy": "workspace:*",
+    "@vitest/utils": "workspace:*",
     "birpc": "^0.2.3",
     "cac": "^6.7.14",
     "chai-subset": "^1.6.0",
