+++ packages/vitest/src/runtime/run.ts
@@ -1,10 +1,9 @@
 import { performance } from 'perf_hooks'
 import limit from 'p-limit'
+import { GLOBAL_EXPECT, getState, setState } from '@vitest/expect'
 import type { BenchTask, Benchmark, BenchmarkResult, File, HookCleanupCallback, HookListener, ResolvedConfig, SequenceHooks, Suite, SuiteHooks, Task, TaskResult, TaskState, Test } from '../types'
 import { vi } from '../integrations/vi'
 import { clearTimeout, createDefer, getFullName, getWorkerState, hasFailed, hasTests, isBrowser, isNode, isRunningInBenchmark, partitionSuiteChildren, setTimeout, shuffle } from '../utils'
-import { getState, setState } from '../integrations/chai/jest-expect'
-import { GLOBAL_EXPECT } from '../integrations/chai/constants'
 import { takeCoverageInsideWorker } from '../integrations/coverage'
 import { getBenchOptions, getFn, getHooks } from './map'
 import { rpc } from './rpc'
