+++ packages/vitest/src/types/chai.ts
@@ -1,63 +1,8 @@
-import type { use as chaiUse } from 'chai'
-
-/**
- * Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- */
-
-import type * as jestMatcherUtils from '../integrations/chai/jest-matcher-utils'
+import type { MatcherState as JestMatcherState } from '@vitest/expect'
 import type SnapshotState from '../integrations/snapshot/port/state'
 import type { VitestEnvironment } from './config'
 
-export type FirstFunctionArgument<T> = T extends (arg: infer A) => unknown ? A : never
-export type ChaiPlugin = FirstFunctionArgument<typeof chaiUse>
-
-export type Tester = (a: any, b: any) => boolean | undefined
-
-export interface MatcherState {
-  assertionCalls: number
-  currentTestName?: string
-  dontThrow?: () => void
-  error?: Error
-  equals: (
-    a: unknown,
-    b: unknown,
-    customTesters?: Array<Tester>,
-    strictCheck?: boolean,
-  ) => boolean
-  expand?: boolean
-  expectedAssertionsNumber?: number | null
-  expectedAssertionsNumberErrorGen?: (() => Error) | null
-  isExpectingAssertions?: boolean
-  isExpectingAssertionsError?: Error | null
-  isNot: boolean
+export interface MatcherState extends JestMatcherState {
   environment: VitestEnvironment
-  promise: string
   snapshotState: SnapshotState
-  suppressedErrors: Array<Error>
-  testPath?: string
-  utils: typeof jestMatcherUtils & {
-    iterableEquality: Tester
-    subsetEquality: Tester
-  }
 }
-
-export interface SyncExpectationResult {
-  pass: boolean
-  message: () => string
-  actual?: any
-  expected?: any
-}
-
-export type AsyncExpectationResult = Promise<SyncExpectationResult>
-
-export type ExpectationResult = SyncExpectationResult | AsyncExpectationResult
-
-export interface RawMatcherFn<T extends MatcherState = MatcherState> {
-  (this: T, received: any, expected: any, options?: any): ExpectationResult
-}
-
-export type MatchersObject<T extends MatcherState = MatcherState> = Record<string, RawMatcherFn<T>>
