+++ tsconfig.json
@@ -16,6 +16,9 @@
     "paths": {
       "@vitest/ws-client": ["./packages/ws-client/src/index.ts"],
       "@vitest/ui": ["./packages/ui/node/index.ts"],
+      "@vitest/utils": ["./packages/utils/src/index.ts"],
+      "@vitest/spy": ["./packages/spy/src/index.ts"],
+      "@vitest/expect": ["./packages/expect/src/index.ts"],
       "@vitest/browser": ["./packages/browser/src/node/index.ts"],
       "#types": ["./packages/vitest/src/index.ts"],
       "~/*": ["./packages/ui/client/*"],
